# [AIRFLOW-5930] Use cached-SQL query building for hot-path queries