--- a/airflow/api/common/experimental/trigger_dag.py
+++ b/airflow/api/common/experimental/trigger_dag.py
@@ -30,7 +30,6 @@ from airflow.utils.state import State
 def _trigger_dag(
         dag_id: str,
         dag_bag: DagBag,
-        dag_run: DagModel,
         run_id: Optional[str],
         conf: Optional[Union[dict, str]],
         execution_date: Optional[datetime],
@@ -40,7 +39,6 @@ def _trigger_dag(
 
     :param dag_id: DAG ID
     :param dag_bag: DAG Bag model
-    :param dag_run: DAG Run model
     :param run_id: ID of the dag_run
     :param conf: configuration
     :param execution_date: date of execution
@@ -71,7 +69,7 @@ def _trigger_dag(
     if not run_id:
         run_id = "manual__{0}".format(execution_date.isoformat())
 
-    dag_run_id = dag_run.find(dag_id=dag_id, run_id=run_id)
+    dag_run_id = DagRun.find(dag_id=dag_id, run_id=run_id).one_or_none()
     if dag_run_id:
         raise DagRunAlreadyExists("Run id {} already exists for dag id {}".format(
             run_id,
@@ -123,10 +121,8 @@ def trigger_dag(
     if dag_model is None:
         raise DagNotFound("Dag id {} not found in DagModel".format(dag_id))
     dagbag = DagBag(dag_folder=dag_model.fileloc)
-    dag_run = DagRun()
     triggers = _trigger_dag(
         dag_id=dag_id,
-        dag_run=dag_run,
         dag_bag=dagbag,
         run_id=run_id,
         conf=conf,