@shlomiken 2019-12-25
> @ashb - you probably on vacation now , but i wanted to see if you can help so i would be able to make progress with this image you have provided us . astronomerio / ap-airflow : 1.10.5-perf-fixes-alpine3.10-onbuild
> the problem is that i first got some module not found errors and now i get those messages .
> `File "/usr/lib/python3.7/site-packages/sqlalchemy/engine/default.py", line 581, in do_execute cursor.execute(statement, parameters) sqlalchemy.exc.ProgrammingError: (psycopg2.ProgrammingError) relation "connection" does not exist LINE 2: FROM connection GROUP BY connection.conn_id ^`
> i really want to test this as soon as possible , but cannot tell what is wrong with this image.
> thanks
> Shlomi

I think that the connection issue is behind me - was related to astro-cli version i guess.
but now i see this one 
`Traceback (most recent call last):
  File "/usr/bin/airflow", line 32, in <module>
    args.func(args)
  File "/usr/lib/python3.7/site-packages/airflow/utils/cli.py", line 74, in wrapper
    return f(*args, **kwargs)
  File "/usr/lib/python3.7/site-packages/airflow/bin/cli.py", line 1017, in scheduler
    job.run()
  File "/usr/lib/python3.7/site-packages/airflow/jobs/base_job.py", line 213, in run
    self._execute()
  File "/usr/lib/python3.7/site-packages/airflow/jobs/scheduler_job.py", line 1317, in _execute
    from airflow.ti_deps.deps.trigger import TriggerRuleDep
ModuleNotFoundError: No module named 'airflow.ti_deps.deps.trigger'`