@ashb 2020-01-11
Hi @shlomiken Sorry, my inbox is a disaster area right now :) I've fixed that issue, and a few more.

Something is still up with the tests here on master that I need to dig in to, but the backported version of this change to 1.10.5 should be ready tomorrow/Monday: https://travis-ci.com/astronomer/airflow/builds/143977776