@ashb 2020-01-15
Tests aren't quite passing -- one failure due to a non-fixed ordering of return values.