@ashb 2020-01-16
Updated numbers from master branch. I'm not sure these will equate to _such_ a speed up in terms of actual task execution. Overall numbers are lower as I have a faster machine to run tests on now :)


  Concurrent DagRuns | Tasks | Before | After | Speedup
  -- | -- | -- | -- | --  
 100 | 12 | 3.9545s (±0.140s) | 1.7174s (±0.045s) | x2.3
 100 | 40 | 12.1224s (±0.357s) |9.0649s (±0.603s) | x1.3
 1000 | 12 | 17.7790s (±0.240s) | 8.3937s (±0.516s) | x1.5
 1000 | 40 | 67.1121s (±0.799s) | 44.6126s (±0.518s) | x1.5