@ashb 2020-02-21
This PR may not be needed so much anymore once PRs such as #7477 land (this would still make _a bit_ of difference, but possibly not as much anymore.)

I will re-investigate this in a week or two.